+++ src/views/pages/Payslip.vue
@@ -397,7 +397,7 @@ export default {
         if (this.choosenDepartment.id == 1) {
           this.$router.push("/viewPayslip").catch(() => {});
         } else if (this.choosenDepartment.id == 2 || this.choosenDepartment.id == 4) {
-          this.$router.push("/viewPayslip").catch(() => {});
+          this.$router.push("/viewPayslipBulananOffice").catch(() => {});
         } else if (this.choosenDepartment.id == 3) {
           this.$router.push("/viewPayslipCs").catch(() => {});
         }