+++ CHANGELOG.md
@@ -231,3 +231,7 @@
 # 0.0.1 (Update 04/04/2023)
 - Modul Payslip
     - Enable Payslip Bulanan dan Office    
+
+# 0.0.1 (Update 05/04/2023)
+- Modul Payslip Bulanan dan Office
+    - format Payslip dan PDF